Remove Repeating checkbox and add Forever and Times
The checkbox called Repeating was just creating one extra step
in order to create expenses. Instead of that now you can set
a the repeatition of the expenses by choosing forever or repeat.
Forever will create expenses for well.. forever. Repeat will enable an
input in which you can choose how many times you want the expense to be
registered.